done task's view feature employee

+++ src/main/java/com/example/Employee_Management_System/service/impl/ManagerServiceImpl.java
@@ -6,15 +6,14 @@ import com.example.Employee_Management_System.dto.request.CreateTaskRequest;
 import com.example.Employee_Management_System.dto.request.UpdateTaskRequest;
 import com.example.Employee_Management_System.dto.response.Response;
 import com.example.Employee_Management_System.dto.response.WorkingScheduleResponse;
+import com.example.Employee_Management_System.repository.EmployeeRepository;
 import com.example.Employee_Management_System.repository.ManagerRepository;
-import com.example.Employee_Management_System.repository.UserRepository;
 
 import com.example.Employee_Management_System.model.ReportBasicInfo;
 import com.example.Employee_Management_System.repository.TaskRepository;
 import com.example.Employee_Management_System.service.ManagerService;
 import com.example.Employee_Management_System.service.ReportService;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -33,6 +32,8 @@ public class ManagerServiceImpl implements ManagerService {
 
     private final ManagerRepository managerRepository;
 
+    private final EmployeeRepository employeeRepository;
+
     private final TaskRepository taskRepository;
 
     private final ReportService reportService;
@@ -40,7 +41,9 @@ public class ManagerServiceImpl implements ManagerService {
     @Override
     public ResponseEntity<Response> createTask(User manager, CreateTaskRequest request) {
         //TODO: (Vu) check if the employeeId is one of the employees of the manager
-        Long employeeId = request.getEmployeeId();
+        if (checkEmployeeBelongsToManager(manager.getId(), request.getEmployeeId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Employee not belong to the manager!");
+        }
 
         Task task = Task
                 .builder()
@@ -69,9 +72,13 @@ public class ManagerServiceImpl implements ManagerService {
     @Override
     public ResponseEntity<Response> deleteTask(User manager, long taskId) {
         //TODO:(Vu) check if the task belongs to an employee of the manager
-        Task task = taskRepository
-                .getTask(taskId)
+        Task task = managerRepository
+                .getTaskById(taskId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found!"));
+
+        if (checkEmployeeBelongsToManager(manager.getId(), task.getEmployeeId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Employee not belong to the manager!");
+        }
 
         taskRepository.deleteTask(task);
         return ResponseEntity.ok(
@@ -86,11 +93,14 @@ public class ManagerServiceImpl implements ManagerService {
     @Override
     public ResponseEntity<Response> updateTask(User manager, long taskId, UpdateTaskRequest updateTaskRequest) {
         //TODO:(Vu) check if the task belongs to an employee of the manager
-
         // Todo: throw custom exception
-        Task task = taskRepository
-                .getTask(taskId)
+        Task task = managerRepository
+                .getTaskById(taskId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found!"));
+
+        if (checkEmployeeBelongsToManager(manager.getId(), task.getEmployeeId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Employee not belong to the manager!");
+        }
 
         task.setTitle(updateTaskRequest.getTitle());
         task.setDescription(updateTaskRequest.getDescription());
@@ -110,6 +120,13 @@ public class ManagerServiceImpl implements ManagerService {
                         .message("Update task successfully!")
                         .build()
         );
+    }
+
+    private boolean checkEmployeeBelongsToManager(Long managerId, Long employeeId) {
+        return employeeRepository
+                .getAllEmployeesByManagerId(managerId)
+                .stream()
+                .noneMatch(employee -> Objects.equals(employee.getId(), employeeId));
     }
